More load fixing
* Fixed so GameName is set after iwadName
* Fix BuildAllNodes to handle cancelled load
* Fixed compile error on POSIX

+++ src/ui_file.cc
@@ -686,7 +686,7 @@ static void PopulateIWADsHelper(Fl_Choice *choice, const SString &prev_game,
 				while (*p && *p != '|')
 					p++;
 
-				SString gameName(start, p - start);
+				SString gameName(start, (int)(p - start));
 
 				// Check if this game is in the filter list
 				bool isValid = false;
